Set the vim-airline symbol for whitespace
Set a specific symbol for vim-airline to use in the statusline when
indicating trailing whitespace.  The default symbol of a 12-pointed star
is not supported by the Terminus font.  My preference is for the symbol
S/P, but Terminus does not support that either.  Settle for the
double-dagger character.  This renders as a set of three 1-pixel lines
in my usual font size.