Add Vim hotkeys for whitespace and tagbar
In Vim, map the F7 key to toggle whitespace detection in the vim-airline
plugin.  (Detecting trailing whitespace is not always needed, and it
sometimes takes up space in the status line.)

Also, map F8 to toggle the presence of the tagbar side window as
suggested in the tagbar readme file.  Additionally, set tagbar to sort
items by file order by default (as opposed to alphabetical order).