Manually force Vim filetype on .gitconfig
Apply the 'gitconfig' file type to the .gitconfig file even when Vim
does not auto-detect.

Vim does not auto-detect the file type when the Git configuration file
is edited under another name, e.g. ~/dotfiles/gitconfig instead of
~/.gitconfig as expected.  Explicitly providing the filetype in the mode
line fixes the issue and allows for the file to be edited from within
the Git repository.  In turn this allows vim-fugitive to work correctly.