Change max line number symbol to supported char
The airline plugin now displays the total number of lines in a file,
alongside a Unicode character of three horizontal lines.  The default
character is from a set of east Asian cultural symbols.  Replace it with
the mathematical "identical to" operator.  The latter is supported by
the Terminus font, whereas the former is not.