Don't commit Vim spelling files
Tell Git to ignore Vim spelling files.  When telling Vim that a word is
acceptable, Vim will store it in a word list in a particular folder.
Since unknown words are likely to be proper nouns and may relate to work
that is being done, they should not be committed to the repository as it
may constitute an information leak.  Ignoring the files means more work,
but each host effectively has its own word list.