Bash:  override any command-not-found utilities
Ubuntu (and other Debian-based distributions) like to try to be
"helpful" and query the package DB if a command isn't recognized.  This
causes some noticeable lag in many circumstances, and can lead to
frustration if typos are happening.  This commit defines a Bash function
to give a quick, basic message instead of anything more in-depth.